Update Ruby to 3.3 and fix dependencies accordingly
Starting with version 3.19.3, html-proofer removes its dependency
on the end-of-life nokogumbo library and replaces it with nokogirl.
https://github.com/gjtorikian/html-proofer/issues/660

Front_matter_parser 1.0.1 fixes an error with newer versios of Ruby
https://github.com/waiting-for-dev/front_matter_parser/blob/main/CHANGELOG.md#101---2021-07-20

The base64 and csv gems will no longer be included with Ruby, so add to Gemfile
to ensure they will be installed

Update bundler to fix up some warnings/errors and update the rest of the packages

Update to Ruby 3.3

Ruby 3.0 will no longer be supported as of 2024-03-31, so let's get
it upgraded.

research.rb and teaching.rb are included in this changeset because
they were using File.exists, which is invalid syntax as of Ruby 3.2

ignore env helper files